Safe wrappers for Gluster
